requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
Django==2.1.7
django-cors-headers==2.4.0
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
factory-boy==2.11.1
entrypoints==0.3
Faker==1.0.2
flake8==3.7.5
httplib2==0.12.0
idna==2.8
isort==4.3.4
mccabe==0.6.1
more-itertools==4.3.0
oauth2==1.9.0.post1
Pillow==5.4.1
pluggy==0.8.0
psycopg2-binary==2.7.7
py==1.7.0
pycodestyle==2.5.0
pyflakes==2.1.0
PyJWT==1.7.1
pytest==3.9.3
pytest-cov==2.6.0
pytest-django==3.4.3
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
text-unidecode==1.2
urllib3==1.24.1
